Use 'yarn' directly for multi-platform support
`$npm_execpath` runs yarn or npm depending on which one you used to run
the script, but since I'm sticking with `yarn` for this project I'll use
just that.

Env variables on Windows use `%` instead of `$`, so the `$npm_execpath`
"trick" doesn't work on Windows, at least how I was using it.